+++ src/Tests/Dfe.Complete.CypressTests/cypress/auth/authenticationInterceptor.ts
@@ -14,6 +14,7 @@ export class AuthenticationInterceptor {
                     ...req.headers,
                     Authorization: `Bearer ${Cypress.env(EnvAuthKey)}`,
                     "x-user-context-name": user.username, // must be present, but not used
+                    "x-user-context-id": user.id, // must be present for antiforgery claims
                     "x-user-ad-id": user.adId,
                     "x-cypress-user": userType,
                 };