+++ CHANGELOG.md
@@ -25,8 +25,11 @@ Security in case of vulnerabilities.
 ### Added
 
 ### Changed  
+- Add a redirect from projects/{id} to projects/{id}/tasks
+- Reduce log level from error to warning when project routes receive bad GUID or project not found 
 
 ### Fixed
+- Remove double pagination on Service Support > Local authorities
 
 See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-07-14.525...main) for everything awaiting release
 
@@ -61,6 +64,8 @@ See the [full commit history](https://github.com/DFE-Digital/complete-conversion
 - New route /projects/{projectId}/academy-urn
 - Service support Create Academy URN (/projects/{projectId}/academy-urn) 
 - Added `Statistics` page
+- Added Project Significant History Dates `/projects/{projectId}/date-history`
+
 
 ### Changed